Create a single CucumberJava8ToCucumberJavaTest with @Nested sub classes
